Split tests into separate namespaces and classes
