+++ src/Reline.Tests/Tests/ParserTests/Parameters.cs
@@ -1,64 +1,9 @@
 ﻿using Reline.Compilation.Syntax;
 using Reline.Compilation.Syntax.Nodes;
 
-namespace Reline.Tests;
+namespace Reline.Tests.ParserTests;
 
-public class ParserTests : ParserTestBase {
-
-	[Fact]
-	public void ParsesProgramStructure() {
-		string source =
-@"
-
-";
-		var (r, _) = Parse(source);
-
-		r.Node<ProgramSyntax>();
-		{
-			r.Node<LineSyntax>();
-			r.Node<LineSyntax>();
-			r.Node<LineSyntax>();
-		}
-		r.End();
-	}
-
-	[Fact]
-	public void ParsesFunctionDeclarations() {
-		string source =
-@"function Foo 2..3 (a b c)
-
-";
-		var (r, _) = Parse(source);
-
-		r.Node<ProgramSyntax>();
-		{
-			// function Foo 2..3 (a b c)
-			r.Node<LineSyntax>();
-			{
-				// function
-				r.Node<FunctionDeclarationStatementSyntax>()
-					.IdentifierIs("Foo");
-				// 2..3
-				{
-					// ..
-					r.Node<BinaryExpressionSyntax>()
-						.OperatorTypeIs(SyntaxType.DotDotToken);
-					// 2
-					r.Node<LiteralExpressionSyntax>()
-						.ValueEquals(2);
-					// 3
-					r.Node<LiteralExpressionSyntax>()
-						.ValueEquals(3);
-				}
-				// (a b c)
-				r.Node<ParameterListSyntax>()
-					.ParametersAre(new[] { "a", "b", "c" });
-			}
-			r.Node<LineSyntax>();
-			r.Node<LineSyntax>();
-		}
-		r.End();
-	}
+public class Parameters : ParserTestBase {
 
 	[Fact]
 	public void ParsesSingleParameterFunctionInvocations() {
@@ -223,135 +168,6 @@ public class ParserTests : ParserTestBase {
 						// d
 						r.Node<IdentifierExpressionSyntax>()
 							.IdentifierIs("d");
-					}
-				}
-			}
-		}
-		r.End();
-	}
-
-	[Fact]
-	public void ParsesBinaryOperators() {
-		string source =
-@"1 + 2 * 3 - 4 .. 5 / 6 % 7
-""a"" < ""b""";
-		var (r, _) = Parse(source);
-
-		r.Node<ProgramSyntax>();
-		{
-			// 1 + 2 * 3 - 4 .. 5 / 6 % 7
-			r.Node<LineSyntax>();
-			{
-				r.Node<ExpressionStatementSyntax>();
-				{
-					// ..
-					r.Node<BinaryExpressionSyntax>()
-						.OperatorTypeIs(SyntaxType.DotDotToken);
-					// 1 + 2 * 3 - 4
-					{
-						// -
-						r.Node<BinaryExpressionSyntax>()
-							.OperatorTypeIs(SyntaxType.MinusToken);
-						// 1 + 2 * 3
-						{
-							// +
-							r.Node<BinaryExpressionSyntax>()
-								.OperatorTypeIs(SyntaxType.PlusToken);
-							{
-								// 1
-								r.Node<LiteralExpressionSyntax>()
-									.ValueEquals(1);
-								// 2 * 3
-								{
-									// *
-									r.Node<BinaryExpressionSyntax>()
-										.OperatorTypeIs(SyntaxType.StarToken);
-									// 2
-									r.Node<LiteralExpressionSyntax>()
-										.ValueEquals(2);
-									// 3
-									r.Node<LiteralExpressionSyntax>()
-										.ValueEquals(3);
-								}
-							}
-						}
-						// 4
-						r.Node<LiteralExpressionSyntax>()
-							.ValueEquals(4);
-					}
-					// 5 / 6 % 7
-					{
-						// %
-						r.Node<BinaryExpressionSyntax>()
-							.OperatorTypeIs(SyntaxType.PercentToken);
-						// 5 / 6
-						{
-							// /
-							r.Node<BinaryExpressionSyntax>()
-								.OperatorTypeIs(SyntaxType.SlashToken);
-							// 5
-							r.Node<LiteralExpressionSyntax>()
-								.ValueEquals(5);
-							// 6
-							r.Node<LiteralExpressionSyntax>()
-								.ValueEquals(6);
-						}
-						// 7
-						r.Node<LiteralExpressionSyntax>()
-							.ValueEquals(7);
-					}
-				}
-			}
-			// "a" < "b"
-			r.Node<LineSyntax>();
-			{
-				r.Node<ExpressionStatementSyntax>();
-				{
-					// <
-					r.Node<BinaryExpressionSyntax>()
-						.OperatorTypeIs(SyntaxType.LesserThanToken);
-					// "a"
-					r.Node<LiteralExpressionSyntax>()
-						.ValueEquals("a");
-					// "b"
-					r.Node<LiteralExpressionSyntax>()
-						.ValueEquals("b");
-				}
-			}
-		}
-		r.End();
-	}
-	[Fact]
-	public void ParsesUnaryOperators() {
-		string source =
-@"+1 + -2";
-		var (r, _) = Parse(source);
-
-		r.Node<ProgramSyntax>();
-		{
-			// +1 + -2
-			r.Node<LineSyntax>();
-			{
-				r.Node<ExpressionStatementSyntax>();
-				{
-					// +
-					r.Node<BinaryExpressionSyntax>()
-						.OperatorTypeIs(SyntaxType.PlusToken);
-					{
-						// +
-						r.Node<UnaryExpressionSyntax>()
-							.OperatorTypeIs(SyntaxType.PlusToken);
-						// 1
-						r.Node<LiteralExpressionSyntax>()
-							.ValueEquals(1);
-					}
-					{
-						// -
-						r.Node<UnaryExpressionSyntax>()
-							.OperatorTypeIs(SyntaxType.MinusToken);
-						// 2
-						r.Node<LiteralExpressionSyntax>()
-							.ValueEquals(2);
 					}
 				}
 			}
